requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
openai>=1.35.10
pydantic>=2.7.0
yfinance>=0.2.40
httpx>=0.27.0
python-dotenv>=1.0.1
tavily-python>=0.3.3
